# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Milk Delivery Customer Management API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/orders": {
      "get": {
        "summary": "List orders",
        "tags": [
          "Orders"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "customerId": {
                            "type": "integer"
                          },
                          "productId": {
                            "type": "integer"
                          },
                          "quantity": {
                            "type": "integer"
                          },
                          "unitPrice": {
                            "type": "integer"
                          },
                          "totalPrice": {
                            "type": "integer"
                          },
                          "orderDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "deliveryDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "status": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "customerId",
                          "productId",
                          "quantity",
                          "unitPrice",
                          "totalPrice",
                          "orderDate",
                          "status",
                          "createdAt"
                        ]
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create order",
        "tags": [
          "Orders"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customerId": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "productId": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "unitPrice": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "orderDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "deliveryDate": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "customerId",
                  "productId",
                  "quantity",
                  "unitPrice",
                  "orderDate"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "customerId": {
                          "type": "integer"
                        },
                        "productId": {
                          "type": "integer"
                        },
                        "quantity": {
                          "type": "integer"
                        },
                        "unitPrice": {
                          "type": "integer"
                        },
                        "totalPrice": {
                          "type": "integer"
                        },
                        "orderDate": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "deliveryDate": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "status": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "customerId",
                        "productId",
                        "quantity",
                        "unitPrice",
                        "totalPrice",
                        "orderDate",
                        "status",
                        "createdAt"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/by-customer/{customerId}": {
      "get": {
        "summary": "List orders by customer and period",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "customerId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "to",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "customerId": {
                            "type": "integer"
                          },
                          "productId": {
                            "type": "integer"
                          },
                          "quantity": {
                            "type": "integer"
                          },
                          "unitPrice": {
                            "type": "integer"
                          },
                          "totalPrice": {
                            "type": "integer"
                          },
                          "orderDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "deliveryDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "status": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "customerId",
                          "productId",
                          "quantity",
                          "unitPrice",
                          "totalPrice",
                          "orderDate",
                          "status",
                          "createdAt"
                        ]
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "put": {
        "summary": "Update order",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customerId": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "productId": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "unitPrice": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "orderDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "deliveryDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "completed",
                      "cancelled"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "customerId": {
                          "type": "integer"
                        },
                        "productId": {
                          "type": "integer"
                        },
                        "quantity": {
                          "type": "integer"
                        },
                        "unitPrice": {
                          "type": "integer"
                        },
                        "totalPrice": {
                          "type": "integer"
                        },
                        "orderDate": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "deliveryDate": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "status": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "customerId",
                        "productId",
                        "quantity",
                        "unitPrice",
                        "totalPrice",
                        "orderDate",
                        "status",
                        "createdAt"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete order",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
...
}